refactor(eidos): extract knowledge::scope submodule (#2642)
Move MemoryScope and ScopeAccessPolicy to knowledge/scope.rs.
All items re-exported from mod.rs.
